+++ src/components/user/index.ts
@@ -11,3 +11,4 @@ export { WishlistButton } from "./WishlistButton";
 export * from "./addresses";
 export * from "./settings";
 export * from "./profile";
+export * from "./notifications";